Migrate coffee details pages and elements to typescript
